Init
This is my personal Yocto "kitchen sink", for creating images which run
on the machines around my house which don't run Debian or Gentoo. It's
a bit messy and too monolithic, but none of it is really specific to me,
so maybe somebody else will find it useful.

Eventually, I need to split this up into multiple layers, and separate
the generic recipes so they are actually reuseable in other layers.